+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import { ExtractComponentController } from "./controllers/ExtractComponentController";
+import { HelpController } from "./controllers/HelpController";
 
 export function activate(context: vscode.ExtensionContext) {
   console.log("react-refactor-to-file extension is now active");
@@ -20,8 +21,18 @@ export function activate(context: vscode.ExtensionContext) {
     }
   );
 
+  const helpController = new HelpController();
+
+  const disposableHelp = vscode.commands.registerCommand(
+    "react-refactor-to-file.help",
+    () => {
+      helpController.showHelp();
+    }
+  );
+
   context.subscriptions.push(disposable);
   context.subscriptions.push(disposableCustom);
+  context.subscriptions.push(disposableHelp);
 }
 
 export function deactivate() {}